fix(job-query): 修复分页偏移量溢出，排序改为 created_at DESC, id DESC (#43)
- page * size 改用 long 计算，offset > 10000 返回 400 PAGE_OUT_OF_RANGE
- 排序改为 created_at DESC, id DESC，同一时间的记录顺序固定
- 新增 JobQueryServiceTests（溢出/回绕/边界）与 JobDescriptionMapperMySqlTests（真实 MySQL 验证排序与翻页不重不漏）

+++ jd-rag-resume-backend/src/main/java/com/arthur/jdragresume/mapper/JobDescriptionMapper.java
@@ -5,11 +5,12 @@ import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
+
 @Mapper
 public interface JobDescriptionMapper {
     List<JobListItem> searchJobs(@Param("userId") Long userId,
                                  @Param("keyword") String keyword,
                                  @Param("location") String location,
-                                 @Param("offset") int offset,
+                                 @Param("offset") long offset,
                                  @Param("size") int size);
 }
